Allow users to sign up for waitlist for activities
This will allow users to sign up for activities that are full and then be automatically inserted into an activity when a spot opens up. The order of the waitlist is determined by the time the user joins the waitlist, with preference given to the user that signed up first. This only comes into effect once the activity is full, so this will not cause problems with relation to 48-hr signups.

TODO: Give user information about their waitlisted activities and properly change the UI to display the correct information that is changed via JS. We could also possibly send the user an email when they are signed up for an activity they are on the waitlist for.

Signed-off-by: Omkar Kulkarni <2019okulkarn@tjhsst.edu>