refactor(dashboard): fix hanging indent in link widgets
Currently, the way margins are applied to the link widgets means that if
the link text wraps around, the second line is unindented relative to
the first line. This makes the it so that the second line is indented.

As part of doing this, I had to make the .link-headings block elements.
I used <h5>s because they *are* headings, and <h5>s render with 100%
text size.