fix(eighth): don't display pie-0 icon for activities with signups
Currently, if fewer users than 10% of the capacity are signed up, the
activity is shown with the pie-0 icon. This incorrectly implies that the
activity has no signups, when it actually does.
This commit changes that behavior so that the pie-0 icon will never be
shown for activities with at least one user signed up.