fix(hoco): position hoco scores as absoluute so scrollbars don't appear
The login form is positioned such that there's enough padding above it.

This also actually fixes a bug where the events sidebar and the
homecoming scores on the login page interfere with each other (i.e. the
login page is slightly broken when there is an event during homecoming
week. This bug was discovered about two years ago, but it would appear
that nobody has fixed it. <shrug> I guess I get to be the luck one.