fix(schedule): remove duplicated JS includes
This fixes a bug where the dropdown in the nav does not work on the
calendar page. Essentially, we use a custom animation "easing" function
that's defined in jQuery UI, but the re-include of jQuery (core)
replaces the easing function list, removing that function.

Also, in general we shouldn't be including the same JS file twice.