refactor: refactor Redis settings
The settings for django-redis-sessions are not specified properly.

Additionally, the use of the virtual environment name in the Redis
prefix is unnecessary and can only cause problems later on. This seems
to have been intended to allow for testing and production environments
on the same server, but that is not necessary with the current
production/testing environment setup.