perf(announcements): simplify Announcement.is_visible_requester query
WARNING: This does change behaviour slightly in the event that multiple
AnnouncementRequests exist for a given Announcement. However, I do not
see a problem for this as this is an edge case that should never
occur (and I have confirmed that this is not the case for any
Announcement in production).