docs(postinstall): remove ambiguity in the initial setup instructions
In the paragraph starting with "You can then start...", these instructions tell the reader to login with the master password. However, this is before telling the reader is informed on how to create a user that they can log in with, creating ambiguity. This edit attempts to make it more clear what to do.