fix(dashboard): limit time when senior graduation message is shown
Currently, unless commented manually, the senior forwarding message on
the dashboard is rendered for all seniors, regardless of when it is
during the school year. We intend for the message to be only rendered
near the end of academic year.

This commit restricts the rendering of this message to
settings.NEAR_GRADUATION_DAYS (50).